fix Exception rescue
I am investigating the PayrollValidation job on Jenkins, and in some cases, the `Exception for record in...` message is strange as the validation for the record passes.

I would expect this PR to makes things better and this is a better practice so nothing to lose - see http://stackoverflow.com/questions/10048173/why-is-it-bad-style-to-rescue-exception-e-in-ruby for more information about this change.

🌴

Closes https://github.com/Gusto/broken_record/pull/5